Refactor: Rename arbitratorregistration package to arbitrator
